move reporter types to an intermediate .ts file, to be combined with reporter.js when migrated
